requirements.txt
setuptools==77.0.3
wheel==0.45.1
mypy==1.15.0
typing-extensions==4.12.2
Pillow==11.1.0
google-generativeai==0.8.4
google-genai==1.7.0
openai==1.68.2
python-dotenv==1.0.1
SQLAlchemy==2.0.39
requests==2.32.3
types-requests==2.32.0.20250306
SpeechRecognition==3.14.2
Django==5.0.2
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
python-dotenv==1.0.1
groq==0.4.2
numpy==2.2.5
pydub==0.25.1
pedalboard==0.9.16
google-cloud==0.34.0
google-cloud-texttospeech==2.26.0
transformers==4.51.3
torch==2.7.0
accelerate==1.6.0

